add description and cleanup modal exports
