+++ packages/libs/react-ui/src/index.ts
@@ -1,7 +1,5 @@
 import './styles/global.css';
 
-// TODO remove comment
-
 export type {
   IAccordionProps,
   IAccordionSectionProps,
@@ -74,7 +72,6 @@ export {
   Label,
   Link,
   MaskedValue,
-  Modal,
   ModalProvider,
   NavFooter,
   NavHeader,
